Adds model.query documentation and allows model.query to be a callable
Documentation on ``model.query`` was missing so I decided to write at least a short
paragraph about it. The callable functionally is required in one of my projects
so I can filter the model ressources using a rather "complex" SQLAlchemy query
with multiple joins which can't be easily translated into the search_params-"Language".